Stuff: Support 64 bit integers in my_log2().
Fixes some issues with random() on values larger than 32 bits.

+++ src/stuff.h
@@ -10,7 +10,7 @@
 #include "global.h"
 
 /* Prototypes begin here */
-PMOD_EXPORT int my_log2(size_t x) ATTRIBUTE((const));
+PMOD_EXPORT int my_log2(UINT64 x) ATTRIBUTE((const));
 PMOD_EXPORT double my_strtod(char *nptr, char **endptr);
 
 unsigned INT32 find_next_power(unsigned INT32 x) ATTRIBUTE((const));
